do not query /proc for darwin
