+++ components/fd/get_limit_darwin.go
@@ -11,3 +11,7 @@ func checkFDLimitSupported() bool {
 func getLimit() (uint64, error) {
 	return 0, nil
 }
+
+func getUsage() (uint64, error) {
+	return 0, nil
+}
