+++ components/fd/get_limit_windows.go
@@ -10,3 +10,7 @@ func checkFDLimitSupported() bool {
 func getLimit() (uint64, error) {
 	return 0, nil
 }
+
+func getUsage() (uint64, error) {
+	return 0, nil
+}
